fpl_agent/pi/requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
playwright>=1.40.0
